agno/requirements.txt
agno
python-dotenv
google-genai
duckduckgo-search
pypdf
lancedb
streamlit
pandas
tantivy
yfinance
groq